requirements.txt
dtw-python==1.1.6
editdistance==0.5.3
matplotlib==3.3.2
numpy==1.19.1
onnxruntime==1.8.1
pandas==1.1.3
scipy==1.5.2
seaborn==0.11.0
SoundFile==0.10.2
torch==1.6.0
tqdm==4.50.2
transformers==4.17.0
natsort==7.1.1
POT==0.7.0
rpy2==3.4.2